Conflicts:
	app/src/main/java/com/inspius/canyon/yo_video/api/RPC.java
	app/src/main/java/com/inspius/canyon/yo_video/app/AppConstant.java

+++ app/src/main/java/com/inspius/canyon/yo_video/model/VideoModel.java
@@ -114,7 +114,7 @@ public class VideoModel implements Serializable {
         return videoType;
     }
 
-    public long getVideoId() {
+    public int getVideoId() {
         return videoJSON.id;
     }
 }
